fix all problems

+++ services/userService.js
@@ -1,14 +1,6 @@
 import { userRepository } from "../repositories/userRepository.js";
 
 class UserService {
-  getAll() {
-    const users = userRepository.getAll();
-    if (!users) {
-      return null;
-    }
-    return users;
-  }
-
   search(search) {
     const item = userRepository.getOne(search);
     if (!item) {
@@ -17,17 +9,34 @@ class UserService {
     return item;
   }
 
+  getAll() {
+    const users = userRepository.getAll();
+    if (!users) {
+      throw new Error("Can't find users!");
+    }
+    return users;
+  }
+
+  getOne(id) {
+        const user = this.search({ id: id });
+    if (!user) {
+      throw new Error("Can't find user!");
+    }
+    return user;
+
+  }
+
   create(data) {
     const { email, phoneNumber } = data;
     const userExists =
       this.search({ email: email }) ||
       this.search({ phoneNumber: phoneNumber });
     if (userExists) {
-      return "This email or phone number exists!";
+      throw new Error ("This email or phone number exists!");
     }
     const newUser = userRepository.create(data);
     if (!newUser) {
-      return "Can't create user!";
+      throw new Error ("Can't create user!");
     }
     return newUser;
   }
@@ -37,20 +46,26 @@ class UserService {
     const userExists =
       this.search({ email: email }) ||
       this.search({ phoneNumber: phoneNumber });
+
     if (userExists) {
-      return "This email or phone number exists!";
+      throw new Error ("This email or phone number exists!");
+    }
+    const user = this.search({ id: id });
+    if (!user) {
+      throw new Error ("User not found!");
     }
     const updateUser = userRepository.update(id, data);
-    if (!updateUser) {
-      return "User not found!";
-    }
     return updateUser;
   }
 
   delete(id) {
+    const userForDelete = this.search({ id: id });
+    if (!userForDelete) {
+      throw new Error ("Can't find user!");
+    }
     const deleteUser = userRepository.delete(id);
     if (!deleteUser) {
-      return null;
+      throw new Error ("Can't find user!");
     }
     return deleteUser;
   }
